% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hour.r
\name{hour}
\alias{hour}
\title{Entrega la hora del día a partir de una fecha en formato posix}
\usage{
hour(date)
}
\arguments{
\item{date}{Fecha en formato POSIXt}
}
\value{
hora del dia
}
\description{
Esta función entrega la hora del día como factor
}
\keyword{hour}

