% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datefuncs.r
\name{toPOSIX}
\alias{toPOSIX}
\title{Transforma fecha en string a posix}
\usage{
toPOSIX(date, dateformat = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{date}{Fecha en formato string}

\item{dateformat}{Formato de fecha en string}
}
\value{
fecha en formato POSIXct
}
\description{
Transforma string con formato especificado a POSIXct
}
\keyword{datetime}
\keyword{posix}
\keyword{string}

