% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dow.r
\name{dow}
\alias{dow}
\title{Agrega día de la semana a dataframe a partir de una fecha posix}
\usage{
dow(df, fecha)
}
\arguments{
\item{df}{Dataframe al que agregar la nueva columna}

\item{fecha}{Columna con fechas en formato POSIXt}
}
\description{
Esta función agrega una nueva columna con los días de la semana a un dataframe
}
\keyword{dow}

