% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dow.r
\name{dow}
\alias{dow}
\title{Entrega día de la semana a partir de una fecha posix}
\usage{
dow(date, days = c("Monday", "Tuesday", "Wednesday", "Thursday", "Friday",
  "Saturday", "Sunday"), daylabels = c("lunes", "martes", "miércoles",
  "jueves", "viernes", "sabado", "domingo"))
}
\arguments{
\item{date}{Fecha en formato POSIXt}

\item{days}{Orden para días de la semana}

\item{daylabels}{Formato para días de la semana}
}
\value{
dia de la semana como factor
}
\description{
Esta función entrega el día de la semana en el formato indicado
}
\keyword{dow}

